Finished swapping numbers and moved tasks into packages.
