+++ src/Tara_tasks/OddOrEven/IdentifyIfOddOrEven1.java
@@ -1,4 +1,4 @@
-package Tara_tasks;
+package Tara_tasks.OddOrEven;
 
 
 public class IdentifyIfOddOrEven1 {
